+++ src/parts/await.js
@@ -45,6 +45,7 @@ export function makeAwaitBlock(node, element) {
 
     this.detectDependency(exp);
     this.require('apply', '$cd');
+    if(this.script.readOnly) this.warning('script read-only conflicts with await');
 
     return xNode('await', {
         el: element.bindName(),