+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * style templates, e.g. style:color={color}, style:border-color="red", style:border="1px solid {color}"
 * compile option "css"
 * event modifier: stop, prevent
+* script option "read-only"
 
 ## 0.5.x